+++ example/app/lib/ui/payment_screen.dart
@@ -1,6 +1,7 @@
 import 'package:app/network/network_service.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:stripe_sdk/stripe_sdk.dart';
 
 import '../locator.dart';
 
@@ -14,16 +15,54 @@ class PaymentScreen extends StatelessWidget {
       body: ListView(children: <Widget>[
         Card(
           child: ListTile(
-            title: Text('Perform Charge (VISA, Payment Intent)'),
-            onTap: () => createPaymentIntent(context),
+            title: Text('Automatic payment (VISA, Payment Intent)'),
+            onTap: () => createAutomaticPaymentIntent(context),
+          ),
+        ),
+        Card(
+          child: ListTile(
+            title: Text('Manual payment (VISA, Payment Intent)'),
+            onTap: () => createAutomaticPaymentIntent(context),
           ),
         ),
       ]),
     );
   }
 
-  void createPaymentIntent(BuildContext context) async {
+  void createAutomaticPaymentIntent(BuildContext context) async {
     final NetworkService networkService = locator.get();
-    await networkService.createPaymentIntent(10000, "pm_card_visa");
+    final response = await networkService.createAutomaticPaymentIntent(10000);
+    if (response['status'] == "succeeded") {
+      // TODO: success
+      debugPrint("Success before authentication.");
+      return;
+    }
+    final result = await Stripe.instance
+        .confirmPayment(response['clientSecret'], paymentMethodId: "pm_card_threeDSecure2Required");
+    if (result['status'] == "succeeded") {
+      // TODO: success
+      debugPrint("Success after authentication.");
+      return;
+    } else {
+      debugPrint("Error");
+    }
+  }
+
+  void createManualPaymentIntent(BuildContext context) async {
+    final NetworkService networkService = locator.get();
+    final response = await networkService.createManualPaymentIntent(10000, "pm_card_threeDSecure2Required");
+    if (response['status'] == "succeeded") {
+      // TODO: success
+      debugPrint("Success before authentication.");
+      return;
+    }
+    final result = await Stripe.instance.authenticatePayment(response['clientSecret']);
+    if (result['status'] == "requires_confirmation") {
+      // TODO: make call to server to confirm
+      debugPrint("Success after authentication.");
+      return;
+    } else {
+      debugPrint("Error");
+    }
   }
 }